Update "deleted" management files
