+++ old-management/inventory.py
@@ -6,7 +6,7 @@ import sqlite3
 conn = sqlite3.connect(config.general_database)
 c = conn.cursor()
 
-def take_item(user_id,item,amount=-1):
+def give_item(user_id,item,amount=-1):
     try:
         number = int(item)
     except ValueError:
